# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Club Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /presences:
    get:
      operationId: getPresences
      tags: [presences]
      summary: Get presences
      parameters:
        - name: username
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of presences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Presence"
    post:
      operationId: createPresence
      tags: [presences]
      summary: Record a presence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresenceInput"
      responses:
        "201":
          description: Created

  # Events
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
        "500":
          description: Failed to generate upload URL.

  # Gallery
...
components:
  schemas:
...
    Presence:
      type: object
      required: [id, date, seanceId, categorie, studentUsername, status]
      properties:
        id:
          type: string
        date:
          type: string
        seanceId:
          type: string
        categorie:
          type: string
        studentUsername:
          type: string
        status:
          type: string
    PresenceInput:
      type: object
      required: [date, seanceId, categorie, studentUsername, status]
      properties:
        date:
          type: string
        seanceId:
          type: string
        categorie:
          type: string
        studentUsername:
          type: string
        status:
          type: string
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"